Fix for problems with very long -f filters
We cannot create files with names > 255 on linux, so createing Lockfile with a very
long filter fails. Also, we need exclusion on each job separately. Hence, split up
the filter into each individual jobname and create a lockfile for each. Take care
to release lock we thus acquired on exit.